decimal to integer refactoring

+++ InvestorDashboard.Backend/Services/ITokenService.cs
@@ -7,6 +7,6 @@ namespace InvestorDashboard.Backend.Services
     {
         Task RefreshTokenBalance(string userId = null);
         Task<bool> IsUserEligibleForTransfer(string userId);
-        Task<(string Hash, bool Success)> Transfer(string userId, string destinationAddress, decimal amount);
+        Task<(string Hash, bool Success)> Transfer(string userId, string destinationAddress, long amount);
     }
 }
